+++ ocrd_models/ocrd_models/ocrd_page_generateds.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 #
-# Generated Wed Dec 30 18:18:58 2020 by generateDS.py version 2.35.20.
+# Generated Mon Jan  4 18:05:36 2021 by generateDS.py version 2.35.20.
 # Python 3.6.9 (default, Oct  8 2020, 12:12:24)  [GCC 8.4.0]
 #
 # Command line options:
@@ -3087,7 +3087,7 @@ class PageType(GeneratedsSuper):
         """
         self.invalidate_AlternativeImage(feature_selector='cropped')
         self.Border = Border
-    def get_AllTextLine(self):
+    def get_AllTextLines(self):
         """
         Return all the TextLine in the document
         """