+++ ocrd_models/ocrd_page_user_methods/get_AllTextLines.py
@@ -1,4 +1,4 @@
-def get_AllTextLine(self):
+def get_AllTextLines(self):
     """
     Return all the TextLine in the document
     """